requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
dnspython==2.3.0
fastapi==0.92.0
idna==3.4
Pillow==9.4.0
pydantic==1.10.5
pymongo==4.3.3
python-multipart==0.0.6
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
unicorn==2.0.1.post1
uvicorn==0.15.0
